Changing the less than sign to a greater than sign fixed our problems with the motors getting stuck after hitting the limit.
While using the joysticks to control the arms our momentum will throw the arm slightly past the set limits, so if we are going to use joysticks to control our arms later on we should change that.

We added in code to work the servo. Now we have control of all three axis.

We started working on a method that will control the motors and move them to set positions. This method has a bug, it will not hold its position after it reaches its destination. We had to add a wait after setting a new motor mode.